More work on organizing server code
